requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
torch==2.6.0
torchvision==0.21.0
Pillow==11.2.1
python-multipart